исправил замечания #3
